Fix typos, make exception more helpful if decoding fails
